Fixed initialization issues when script quits on unset variables.
I was testing a number of variables to see if they were null strings to see whether or not they were set. Unfortunately, this meant that the script quit whenever I tested these conditions because testing whether a variable is equal to a null string was checking an unset variable. I fixed the issue by initializing these variables to null strings.